First version. Mayby works - needs to check

+++ server/warehouse/views.py
@@ -1,9 +1,8 @@
 import datetime
-import json
 
 import django
 from django.core.exceptions import ObjectDoesNotExist
-from oauth2_provider.models import AbstractAccessToken, AccessToken
+from oauth2_provider.models import AccessToken
 from rest_framework import status, permissions
 from rest_framework.response import Response
 from rest_framework.status import HTTP_300_MULTIPLE_CHOICES
@@ -15,6 +14,23 @@ from rest_framework.authtoken.models import Token
 from warehouse.models import *
 from warehouse.serializers import ProductSerializer, SyncSerializer
 
+
+def _is_products_same(prod1, prod2):
+    return all(
+        (
+            prod1.man_name == prod2.man_name,
+            prod1.model_name == prod2.model_name,
+            prod1.price == prod2.price,
+            prod1.quantity == prod2.quantity
+         )
+    )
+
+def _end(msg):
+    err = {
+        'error_msg': msg
+    }
+    print(err)
+    return Response(err, status=HTTP_300_MULTIPLE_CHOICES)
 
 class IsManagerOrEmployee(permissions.BasePermission):
 
@@ -41,9 +57,14 @@ class ProductsList(APIView):
 
     def post(self, request):
         serializer = ProductSerializer(data=request.data)
+
         if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            try:
+                Product.objects.get(man_name=serializer.validated_data['man_name'])
+            except Product.DoesNotExist:
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response({'error': 'Product already exists'}, status=status.HTTP_302_FOUND)
         return Response(serializer.errors, status=status.HTTP_201_CREATED)
 
 
@@ -61,27 +82,66 @@ class ProductsDetail(APIView):
     def put(self, request, pk, format=None):
         product = self._get_product(pk)
         if product is None:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        print(request.data)
-        serializer = ProductSerializer(product, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return _end(f'There is no product. Consider sync with server.')
+
+        changes = request.data['new']
+        for prod in request.data['old']:
+            if prod['man_name'] == request.data['new']['man_name']:
+                old = prod
+                break
+        error_fields = []
+        for atr in changes:
+            if self._can_modify(atr, product, old):
+                if atr =='quantity':
+                    product.quantity = old['quantity'] + changes['quantity']
+                    if old['quantity'] + changes['quantity'] < 0:
+                        return _end(f'Quantity cannot be less than 0.')
+                if atr == 'price':
+                    product.price = old['price'] + changes['price']
+                if atr == 'model_name':
+                    product.model_name = changes['model_name']
+            else:
+                error_fields.append(atr)
+        if error_fields:
+            return _end(f'Someone has modified "{error_fields}". You should sync with server')
+        else:
+            product.save()
+
+        return Response({}, status=status.HTTP_200_OK)
 
     def delete(self, request, pk, format=None):
         product = self._get_product(pk)
         if product is None:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+            return _end(f'There is no product. Consider sync with server.')
+        print(request.data)
 
-        product.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        changes = request.data['to_remove']
+        for prod in request.data['old']:
+            if prod['man_name'] == request.data['to_remove']['man_name']:
+                old = prod
+                break
+        error_fields = []
+        for atr in changes:
+            if not self._can_modify(atr, product, old):
+                error_fields.append(atr)
+        if error_fields:
+            return _end(f'Someone has modified "{error_fields}". You should sync with server')
+        else:
+            product.delete()
+
+        return Response({}, status=status.HTTP_200_OK)
 
     def _get_product(self, id):
         try:
             return Product.objects.get(id=id)
         except Product.DoesNotExist:
             return None
+
+    def _can_modify(self, param_name, product, old):
+        if getattr(product, param_name) == getattr(Product(**old), param_name):
+            return True
+        else:
+            return False
 
 
 class AuthView(APIView):
@@ -121,24 +181,69 @@ class SyncView(APIView):
         serializer = SyncSerializer(data=seriable)
         if serializer.is_valid():
             # print(json.dumps(serializer.data, indent=4))
-            # self._mod_prod_handle(serializer.data)
-            self._add_prod_handle(serializer.data)
-            self._rm_prod_handle(serializer.data)
+            err = self._add_prod_handle(serializer.data)
+            if err is not None:
+                return Response(err, status=status.HTTP_300_MULTIPLE_CHOICES)
+            err = self._rm_prod_handle(serializer.data)
+            if err is not None:
+                return Response(err, status=status.HTTP_300_MULTIPLE_CHOICES)
+            err = self._mod_prod_handle(serializer.data)
+            if err is not None:
+                return Response(err, status=status.HTTP_300_MULTIPLE_CHOICES)
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def _mod_prod_handle(self, req_prods):
-        for prod in req_prods['new']:
-            db_prod = Product.objects.get(man_name=prod['man_name'])
-            if db_prod.quantity + prod['quantity'] < 0:
-                err = {
-                    'error_msg': f'Cannot change quantity. You should sync with server'
-                }
-                return Response(err, status=HTTP_300_MULTIPLE_CHOICES)
-            else:
-                if prod['quantity'] != 0:
-                    db_prod.quantity = db_prod.quantity + prod['quantity']
-                    db_prod.save()
+        # 0-> new , 1->old
+        for prod in self._match_old_to_new(req_prods):
+            # Check if product is modified
+            if not _is_products_same(Product(**prod[1]), Product(**prod[0])):
+                prod_name = prod[0]["man_name"]
+                try:
+                    db_prod = Product.objects.get(man_name=prod[0]['man_name'])
+                except Product.DoesNotExist:
+                    err = {
+                        'error_msg': f'Product {prod_name} does not exist any more. You should sync with server'
+                    }
+                    print(err)
+                    return err
+
+                if db_prod.quantity + prod[0]['quantity'] < 0:
+                    err = {
+                        'error_msg': f'Cannot change quantity of {prod_name}. You should sync with server'
+                    }
+                    return err
+                else:
+                    # Check if product is modified.
+                    new_prod = Product(**prod[0])
+                    old_prod = Product(**prod[1])
+                    errs = []
+                    for attr in ("quantity", "price"):
+                        # Check if attribute was changed
+                        print(attr, getattr(new_prod, attr), getattr(old_prod, attr))
+                        if getattr(new_prod, attr) != getattr(old_prod, attr):
+                            # Yes, check if actual value is due to database
+                            if getattr(old_prod, attr) == getattr(db_prod, attr):
+                                # Yes
+                                setattr(db_prod, attr, getattr(new_prod, attr))
+                            else:
+                                errs.append(attr)
+                    if errs:
+                        err = {
+                            'error_msg': f'Cannot make changes in {errs} for product {new_prod.man_name}. You should sync with server'
+                        }
+                        print(err)
+                        return err
+                    else:
+                        db_prod.save()
+
+    def _match_old_to_new(self, req_prods):
+        prods = []
+        for new in req_prods['new']:
+            for old in req_prods['old']:
+                if old['man_name'] == new['man_name']:
+                    prods.append((new, old))
+        return prods
 
     def _add_prod_handle(self, req_prods):
         for prod in req_prods['new']:
@@ -154,7 +259,7 @@ class SyncView(APIView):
                         'error_msg': f'Someone has also added "{prod["man_name"]}". You should sync with server'
                     }
                     print(err)
-                    return Response(err, status=HTTP_300_MULTIPLE_CHOICES)
+                    return err
 
     def _rm_prod_handle(self, req_prods):
         # for diff in list(dictdiffer.diff(req_prods["old"], req_prods["new"])):
@@ -166,7 +271,12 @@ class SyncView(APIView):
         for item in objects_to_remove:
             db_prod = Product.objects.get(man_name=item['man_name'])
             tmp_prod = Product(**item)
-            if tmp_prod == db_prod:
-                print(True)
+            if _is_products_same(tmp_prod, db_prod):
+                db_prod.delete()
+
             else:
-                print(False)
+                err = {
+                    'error_msg': f'Someone has modified "{db_prod.man_name}". You should sync with server'
+                }
+                print(err)
+                return err
